Move registerHook to script file
