+++ all-clients/encode-errorcodes.mjs
@@ -1,8 +1,26 @@
+import { registerHooks } from "node:module";
 import pkg from "./package.json" with { type: "json" };
 import { writeFile } from "node:fs/promises";
 import { restoreErrorCodes } from "./restore-errorcodes.mjs";
 import assert from "node:assert";
-import { gt } from "semver";
+import { gt, satisfies } from "semver";
+
+registerHooks({
+  resolve(specifier, context, nextResolve) {
+    const regex = /^@apollo-client\/(.*?)\/invariantErrorCodes\.js$/;
+    const match = regex.exec(specifier);
+    const version = match?.[1];
+
+    if (!version || !satisfies(version, ">=4.0.0-alpha <=4.0.2")) {
+      return nextResolve(specifier, context);
+    }
+
+    return nextResolve(
+      `./node_modules/@apollo-client/${version}/invariantErrorCodes.js`,
+      context
+    );
+  },
+});
 
 function getLookupArray() {
   const map = new Map();
